+++ src/Produtos/Produtos.java
@@ -5,12 +5,12 @@ public class Produtos {
     private int id;
     private String nome;
     private String tipo;
-    private int preco;       // preço acho melhor padronizarmos em reais, no banco de dados ele aceita o formato: 99.99
+    private float preco;       // preço acho melhor padronizarmos em reais, no banco de dados ele aceita o formato: 99.99
     private long quantidade; //Acho melhor padronizarmos a quantidade em gramas
 
     public Produtos() {}
 
-    public Produtos(int id, String nome, String tipo, int preco, long quantidade) {
+    public Produtos(int id, String nome, String tipo, float preco, long quantidade) {
         this.id = id;
         this.nome = nome;
         this.tipo = tipo;
@@ -28,8 +28,8 @@ public class Produtos {
     public String getTipo() { return tipo; }
     public void setTipo(String tipo) { this.tipo = tipo; }
 
-    public int getPreco() { return preco; }
-    public void setPreco(int preco) { this.preco = preco; }
+    public float getPreco() { return preco; }
+    public void setPreco(float preco) { this.preco = preco; }
 
     public long getQuantidade() { return quantidade; }
     public void setQuantidade(long quantidade) { this.quantidade = quantidade; }
